+++ java/src/main/java/com/google/api/services/samples/authorizedbuyers/marketplace/v1/buyers/clients/ListClients.java
@@ -32,7 +32,7 @@ import net.sourceforge.argparse4j.inf.Namespace;
 public class ListClients {
 
   public static void execute(AuthorizedBuyersMarketplace marketplaceClient, Namespace parsedArgs) {
-    Integer accountId = parsedArgs.getInt("account_id");
+    Long accountId = parsedArgs.getLong("account_id");
     Integer pageSize = parsedArgs.getInt("page_size");
     String parentBuyerName = String.format("buyers/%d", accountId);
     String pageToken = null;
@@ -82,7 +82,7 @@ public class ListClients {
                 + "This will be used to construct the parent used as a path parameter for the "
                 + "clients.list request.")
         .required(true)
-        .type(Integer.class);
+        .type(Long.class);
     parser
         .addArgument("-f", "--filter")
         .help(
